+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.6.14
 
 * Update Translation util
+* Add TextReplacer util to replace any conditional statement by args [like: "Status: {STATUS == active ? \"activated!\" : \"canceled!\"}"]
 
 ## 0.6.13